Further refactoring of cockpit code
Remove cockpit_generic and rename GameCockpit into Cockpit
