+++ engine/src/aldrv/al_sound.cpp
@@ -34,7 +34,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "cmd/unit_generic.h"
-#include "gfx/cockpit_generic.h"
 #include "options.h"
 #include "posh.h"
 #include "universe.h"
