+++ engine/src/cmd/ai/aggressive.cpp
@@ -40,7 +40,6 @@
 #include "flybywire.h"
 #include "hard_coded_scripts.h"
 #include "cmd/script/mission.h"
-#include "gfx/cockpit_generic.h"
 #include "lin_time.h"
 #include "faction_generic.h"
 #include "cmd/role_bitmask.h"
