+++ engine/src/cmd/ai/comm_ai.cpp
@@ -27,7 +27,6 @@
 #include "faction_generic.h"
 #include "communication.h"
 #include "cmd/collection.h"
-#include "gfx/cockpit_generic.h"
 #include "cmd/images.h"
 #include "configxml.h"
 #include "vs_globals.h"
