+++ engine/src/cmd/script/director_generic.cpp
@@ -32,7 +32,6 @@
 #include "cmd/ai/order.h"
 
 #include "configxml.h"
-#include "gfx/cockpit_generic.h"
 
 #include "python/python_class.h"
