+++ engine/src/cmd/script/script_call_order.cpp
@@ -46,7 +46,6 @@
 #include "cmd/ai/flybywire.h"
 #include "cmd/ai/tactics.h"
 #include "cmd/ai/missionscript.h"
-#include "gfx/cockpit_generic.h"
 #include "mission.h"
 #include "easydom.h"
 #include "universe.h"
