+++ engine/src/cmd/unit_generic.cpp
@@ -53,14 +53,12 @@
 #include "cmd/ai/aggressive.h"
 #include "python/python_class.h"
 #include "missile.h"
-#include "gfx/cockpit_generic.h"
 #include "gfx/vsbox.h"
 #include <algorithm>
 #include "cmd/ai/ikarus.h"
 #include "role_bitmask.h"
 #include "unit_const_cache.h"
 #include "gfx/warptrail.h"
-#include "gfx/cockpit_generic.h"
 #include "csv.h"
 #include "vs_random.h"
 #include "galaxy_xml.h"
