+++ engine/src/cmd/unit_util_generic.cpp
@@ -27,14 +27,12 @@
 #include "cmd/unit_util.h"
 #include "configxml.h"
 #include "vs_globals.h"
-#include "gfx/cockpit_generic.h"
 #include "universe_util.h"
 #include "cmd/ai/docking.h"
 #include "savegame.h"
 #include "cmd/planet.h"
 #include "faction_generic.h"
 #include "cmd/ai/fire.h"
-#include "gfx/cockpit_generic.h"
 #include "role_bitmask.h"
 #include "lin_time.h"
 #include "planetary_orbit.h"
