+++ engine/src/faction_util_generic.cpp
@@ -29,7 +29,6 @@
 #include "vs_logging.h"
 #include "config_xml.h"
 #include "vs_globals.h"
-#include "gfx/cockpit_generic.h"
 #include "cmd/unit_generic.h"
 #include "universe.h"
