+++ engine/src/gfx/cockpit.h
@@ -28,7 +28,6 @@
 
 #ifndef _COCKPIT_H_
 #define _COCKPIT_H_
-#include "gfx/cockpit_generic.h"
 #include "gfxlib.h"
 #include "gfxlib_struct.h"
 #include <vector>
@@ -37,6 +36,7 @@
 class TextPlane;
 class VSSprite;
 class Gauge;
+class SoundContainer;
 class Unit;
 class NavigationSystem;
 #include "radar/radar.h"
@@ -45,17 +45,23 @@ class NavigationSystem;
 #include "camera.h"
 #include "physics.h"
 #include "nav/navscreen.h"
+#include "cockpit/cockpit_events.h"
 using namespace XMLSupport;   // DONT PUT THIS ABOVE HEADERS
 #define NUM_CAM CP_NUMVIEWS
+
+#ifdef NETCOMM_WEBCAM
+#define MAXVDUS (13)
+#else
+#define MAXVDUS (12)
+#endif
+
 /**
  * The Cockpit Contains all displayable information about a particular Unit *
  * Gauges are used to indicate analog controls, and some diagital ones
  * The ones starting from KPS are digital with text readout
  */
 
-
-
-class GameCockpit : public Cockpit
+class Cockpit
 {
     Camera    cam[NUM_CAM];
 
@@ -126,7 +132,7 @@ class GameCockpit : public Cockpit
     /// Used to display the arrow pointing to the currently selected target.
     float  projection_limit_x, projection_limit_y;
     float  inv_screen_aspect_ratio; //Precomputed division 1 / g_game.aspect.
-    virtual void SetParent( Unit *unit, const char *filename, const char *unitmodname, const QVector &startloc );
+
     void LoadXML( const char *file );
     void LoadXML( VSFileSystem::VSFile &f );
     void beginElement( const string &name, const AttributeList &attributes );
@@ -138,22 +144,19 @@ class GameCockpit : public Cockpit
 
 
 
-    ///Draws unit gauges
-    void DrawGauges( Unit *un );
+
     ///Trigger scripted events
     void TriggerEvents( Unit *un );
     NavigationSystem ThisNav;
-    //Draw the arrow pointing to the target.
-    void DrawArrowToTarget(const Radar::Sensor&, Unit*);
-    void DrawArrowToTarget(const Radar::Sensor&, Vector LocalCoordinates);
+
     void updateRadar(Unit * un);
 public:
     std::string textMessage;
     bool editingTextMessage;
     std::string autoMessage;
     float autoMessageTime;
-    virtual void setTargetLabel( const string &msg );
-    virtual string getTargetLabel();
+    void setTargetLabel( const string &msg );
+    string getTargetLabel();
     void ReceivedTargetInfo();
     static void NavScreen( const KBData&, KBSTATE k ); //scheherazade
     static string getsoundfile( string filename );
@@ -162,13 +165,16 @@ public:
     int Autopilot( Unit *target );
     ///Restores the view from the IDentity Matrix needed to draw sprites
     void RestoreViewPort();
-    GameCockpit( const char *file, Unit *parent, const std::string &pilotname );
-    ~GameCockpit();
+    Cockpit( const char *file, Unit *parent, const std::string &pilotname );
+    ~Cockpit();
 
     ///Looks up a particular Gauge stat on unit
     float LookupUnitStat( int stat, Unit *target );
+
     ///Loads cockpit info...just as constructor
     void Init( const char *file );
+    void Init( const char *file, bool isDisabled );
+
     ///Draws Cockpit then restores viewport
     void Draw();
     //void Update();//respawns and the like.
@@ -188,6 +194,7 @@ public:
     static void ForceSwitchControl( const KBData &, KBSTATE );
     static void TurretControl( const KBData &, KBSTATE );
     void SetSoundFile( std::string sound );
+    void SetParent( Unit *unit, const char *filename, const char *unitmodname, const QVector &startloc );
     int GetSoundFile()
     {
         return soundfile;
@@ -199,7 +206,7 @@ public:
     {
         return &ThisNav;
     }
-    virtual std::string GetNavSelectedSystem();
+    std::string GetNavSelectedSystem();
     ///Accesses the current camera
     Camera * AccessCamera()
     {
@@ -215,15 +222,15 @@ public:
         cam[currentcamera].UpdateGFX();
     }
 
-    virtual bool SetDrawNavSystem( bool );
-    virtual bool CanDrawNavSystem();
-    virtual bool CheckCommAnimation( Unit *un );
-    virtual void visitSystem( std::string systemName );
+    bool SetDrawNavSystem( bool );
+    bool CanDrawNavSystem();
+    bool CheckCommAnimation( Unit *un );
+    void visitSystem( std::string systemName );
     void AutoLanding();
     void DoAutoLanding(Unit *, Unit *);
 
-    virtual void SetInsidePanYawSpeed( float speed );
-    virtual void SetInsidePanPitchSpeed( float speed );
+    void SetInsidePanYawSpeed( float speed );
+    void SetInsidePanPitchSpeed( float speed );
 
     bool IsPaused() const;
     // Game is paused
@@ -238,6 +245,162 @@ public:
 protected:
     /// Override to use a specific kind of sound implementation
     SoundContainer* soundImpl(const SoundContainer &specs);
+
+/// Cockpit generic variables
+    ///style of current view (chase cam, inside)
+    enum VIEWSTYLE view;
+
+    int   currentcamera;
+    float radar_time;
+    float gauge_time[UnitImages < void > ::NUMGAUGES];
+
+    /// 8 armor vals and 1 for startfuel
+    float StartArmor[9]; //short fix
+
+    ///saved values to compare with current values (might need more for damage)
+    float maxfuel, maxhull;
+
+    ///this is the parent that Cockpit will read data from
+    UnitContainer parent;
+    UnitContainer parentturret;
+    int   unitfaction;
+
+    ///4 views f/r/l/b
+    float shakin;
+
+    ///Video Display Units (may need more than 2 in future)
+    std::string unitmodname;
+
+    ///Color of cockpit default text
+    ///The font that the entire cockpit will use. Currently without color
+    //Gauge *gauges[UnitImages::NUMGAUGES];
+    //0 means no autopilot...positive autopilto in progress
+    float autopilot_time;
+    UnitContainer autopilot_target; //usually null
+
+    /**
+     * two values that represent the adjustment to perspective needed to center teh crosshairs in the perceived view.
+     */
+    float cockpit_offset, viewport_offset;
+
+    float  cockpit_time;
+    bool   ejecting;
+    bool   going_to_dock_screen;
+    int    partial_number_of_attackers;
+
+private:
+    std::vector< std::string > unitfilename;
+    std::vector< std::string > unitsystemname;
+    std::vector< std::string > unitbasename;
+    std::vector< SoundContainer* > sounds;
+
+public:
+    double secondsWithZeroEnergy;
+    int    number_of_attackers;
+    unsigned int retry_dock;
+    double TimeOfLastCollision;
+    char   jumpok;
+    float godliness;
+    std::string communication_choices;
+    float credits; //how much money player has
+    ///How far away chasecam and pan cam is
+    float zoomfactor;
+    class Flightgroup*fg;
+    class StarSystem*activeStarSystem; //used for context switch in Universe
+    class SaveGame*savegame;
+
+/// Cockpit generic functions
+    static void beginElement( void *userData, const XML_Char *name, const XML_Char **atts );
+    static void endElement( void *userData, const XML_Char *name );
+    static std::string MakeBaseName(const Unit *base);
+
+
+    void updateAttackers();
+    static bool tooManyAttackers(); //checks config file and declares if too many folks are attacking any of the players (to avoid expensive tests where unnecessary).
+
+    bool unitInAutoRegion( Unit *un );
+    void RestoreGodliness();
+
+    const std::string& GetUnitFileName(unsigned int which = 0) const;
+    const std::string& GetUnitSystemName(unsigned int which = 0) const;
+    const std::string& GetUnitBaseName(unsigned int which = 0) const;
+
+    void RemoveUnit(unsigned int which = 0);
+    void PackUnitInfo(vector< std::string > &info) const;
+
+    void UnpackUnitInfo(vector< std::string > &info);
+
+    ///Loads cockpit info...just as constructor
+    void recreate( const std::string &pilotname );
+
+    ///Sets owner of this cockpit
+    //unsigned int whichcockpit;//0 is the first player, 1 is the second and so forth
+
+    Unit * GetSaveParent();
+    bool Update(); //respawns and the like. Returns true if starsystem was reloaded
+
+    bool autoInProgress()
+    {
+        return autopilot_time > 0;
+    }
+
+    ///Sets the current viewstyle
+    void SetView( const enum VIEWSTYLE tmp )
+    {
+        view = tmp;
+    }
+    enum VIEWSTYLE GetView()
+    {
+        return view;
+    }
+
+    std::string& GetUnitFileName(unsigned int which = 0)
+    {
+        while ( which >= unitfilename.size() )
+            unitfilename.push_back("");
+        return unitfilename[which];
+    }
+
+    std::string& GetUnitSystemName(unsigned int which = 0)
+    {
+        while ( which >= unitsystemname.size() )
+            unitsystemname.push_back("");
+        return unitsystemname[which];
+    }
+
+    std::string& GetUnitBaseName(unsigned int which = 0)
+    {
+        while ( which >= unitbasename.size() )
+            unitbasename.push_back("");
+        return unitbasename[which];
+    }
+
+    std::string GetUnitModifications()
+    {
+        return unitmodname;
+    }
+
+    size_t GetNumUnits() const
+    {
+        return unitfilename.size();
+    }
+
+    Unit * GetParent()
+    {
+        return parent.GetUnit();
+    }
+
+    /**
+     * Retrieves the sound associated to the given event.
+     * Returns NULL if no sound has been associated
+     */
+    SoundContainer* GetSoundForEvent(CockpitEvent eventId) const;
+
+    /**
+     * Sets the sound associated to the given event to match the given specs
+     * (the actual container will be a different, implementation-specific one)
+     */
+    void SetSoundForEvent(CockpitEvent eventId, const SoundContainer &soundSpecs);
 };
 #endif
