+++ engine/src/gfx/view_style.h
@@ -1,0 +1,10 @@
+#ifndef VIEW_STYLE_H
+#define VIEW_STYLE_H
+
+enum VIEWSTYLE
+{
+    CP_FRONT, CP_BACK, CP_LEFT, CP_RIGHT, CP_VIEWTARGET, CP_PANINSIDE, CP_CHASE, CP_PAN, CP_PANTARGET, CP_TARGET, CP_FIXED,
+    CP_FIXEDPOS, CP_FIXEDPOSTARGET, CP_NUMVIEWS
+};
+
+#endif // VIEW_STYLE_H
