+++ engine/src/savegame.cpp
@@ -38,7 +38,6 @@
 #include "load_mission.h"
 #include <algorithm>
 #include "cmd/script/mission.h"
-#include "gfx/cockpit_generic.h"
 #include "cmd/fg_util.h"
 #include "star_system.h"
 #include "universe.h"
